Move nina coordinator and entity to their own file
